+++ src/visiongui/driver/wait_for_window_to_disappear.py
@@ -12,7 +12,7 @@ logger = logging.getLogger(__name__)
 
 def wait_for_window_to_disappear(
     title: re.Pattern[str],
-    timeout: int,
+    timeout: float,
 ) -> None:
     def _window_check() -> bool:
         return _get_matching_window(title) is None
